+++ Revolver/Controls/Run/quadDirectionalRun.cs
@@ -10,11 +10,11 @@ namespace Revolver.Controls.Run
     internal class quadDirectionalRun : IRun
     {
         public float TimeRunning { get; set; }
-        public int Speed { get; set; }
+        public float Speed { get; set; }
         public float SpeedCap { get; set; }
         public float Acceleration { get; set; }
 
-        public quadDirectionalRun(int Speed, float SpeedCap, float Acceleration)
+        public quadDirectionalRun(float Speed, float SpeedCap, float Acceleration)
         {
             TimeRunning = 0;
             this.Speed = Speed;
